update default handler utils to remove default handler and add defaultdecoderinjector for composition
