+++ pkg/admission/handlers/default.go
@@ -1,7 +1,6 @@
 package handlers
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"net/http"
@@ -14,31 +13,14 @@ type DefaultDecoderInjector struct {
 }
 
 func (d *DefaultDecoderInjector) InjectDecoder(decoder *admission.Decoder) error {
+	if decoder == nil {
+		return fmt.Errorf("decoder cannot be nil")
+	}
 	d.decoder = decoder
 	return nil
 }
 
-type DefaultHandler struct {
-	Decoder *admission.Decoder
-}
-
-func (dh *DefaultHandler) Kind() string {
-	return "default"
-}
-
-func (dh *DefaultHandler) Handle(ctx context.Context, req admission.Request) admission.Response {
-	return admission.Errored(http.StatusBadRequest, fmt.Errorf("resource %s not implemented", dh.Kind()))
-}
-
-func (dh *DefaultHandler) InjectDecoder(d *admission.Decoder) error {
-	if d == nil {
-		return fmt.Errorf("decoder cannot be nil")
-	}
-	dh.Decoder = d
-	return nil
-}
-
-func (dh *DefaultHandler) PatchResponse(raw []byte, v interface{}) admission.Response {
+func PatchResponse(raw []byte, v interface{}) admission.Response {
 	pjson, err := json.Marshal(v)
 	if err != nil {
 		return admission.Errored(http.StatusBadRequest, err)
