+++ pkg/admission/handlers/default_test.go
@@ -1,7 +1,6 @@
 package handlers
 
 import (
-	"context"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -9,38 +8,25 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 )
 
-func TestDefaultHandler_HandleNotImplemented(t *testing.T) {
-	h := &DefaultHandler{}
-	r := h.Handle(context.TODO(), admission.Request{})
-	assert.False(t, r.Allowed)
-}
-
-func TestDefaultHandler_InjectDecoder_FailsOnNil(t *testing.T) {
-	h := &DefaultHandler{}
-	assert.Error(t, h.InjectDecoder(nil))
-}
-
-func TestDefaultHandler_InjectDecoder(t *testing.T) {
-	h := &DefaultHandler{}
+func TestDefaultDecoderInjector(t *testing.T) {
+	d := &DefaultDecoderInjector{}
 	scheme := runtime.NewScheme()
 	decoder, err := admission.NewDecoder(scheme)
 	assert.NoError(t, err)
-	assert.NoError(t, h.InjectDecoder(decoder))
+	assert.NoError(t, d.InjectDecoder(decoder))
 }
 
-func TestDefaultHandler_PatchResponse_ErrorsOnNil(t *testing.T) {
-	h := &DefaultHandler{}
-	resp := h.PatchResponse([]byte{}, "not json")
+func TestPatchResponse_ErrorsOnNil(t *testing.T) {
+	resp := PatchResponse([]byte{}, "not json")
 	assert.Equal(t, false, resp.Allowed)
 }
 
-func TestDefaultHandler_PatchResponse_OK(t *testing.T) {
-	h := &DefaultHandler{}
+func TestPatchResponse_OK(t *testing.T) {
 	d := struct {
 		Hello string
 	}{
 		Hello: "world",
 	}
-	resp := h.PatchResponse([]byte("{}"), &d)
+	resp := PatchResponse([]byte("{}"), &d)
 	assert.Equal(t, true, resp.Allowed)
 }
